requirements.txt
Flask==2.2.5
Flask-CORS==4.0.0
boto3==1.28.25
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0